# src/main/java/com/example/bankcards/dto/request/CardCreateRequest.java
package com.example.bankcards.dto.request;

import jakarta.validation.constraints.Future;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import java.time.LocalDate;

public record CardCreateRequest(
        @NotBlank(message = "Owner username is required")
        String ownerUsername,

        @NotNull(message = "Expiration date is required")
        @Future(message = "Expiration date must be in the future")
        LocalDate expirationDate
) {}